+++ src/components/lodging/LodgingTest.tsx
@@ -1,30 +1,12 @@
 import { useInfiniteScroll } from '@/components/lodging/hooks/useInfiniteScroll';
 import Lodging from '@/components/lodging/Lodging';
 import type { LodgingProps } from '@/components/lodging/types/lodging';
-import React, { useEffect, useRef } from 'react';
+import React, { useRef } from 'react';
 
 export default function LodgingTest() {
   const containerRef = useRef<HTMLUListElement>(null);
   const targetRef = useRef<HTMLDivElement>(null);
-  const { lodging, fetchNextPage, hasNextPage, isFetchingNextPage } = useInfiniteScroll();
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      const observer = new IntersectionObserver(
-        (entries) => {
-          if (entries[0].isIntersecting && hasNextPage && !isFetchingNextPage) {
-            fetchNextPage();
-          }
-        },
-        { root: containerRef.current, threshold: 0.8 }
-      );
-
-      if (targetRef.current) observer.observe(targetRef.current);
-      return () => observer.disconnect();
-    }, 500);
-
-    return () => clearTimeout(timer);
-  }, [targetRef.current, hasNextPage, fetchNextPage]);
+  const { lodging } = useInfiniteScroll({ containerRef, targetRef });
 
   return (
     <div className="w-full h-full flex justify-center items-center bg-amber-100 overflow-hidden">
