+++ src/components/lodging/hooks/useInfiniteScroll.ts
@@ -1,4 +1,10 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
+import { useEffect } from 'react';
+
+type UseInfiniteScrollProps = {
+  containerRef: React.RefObject<HTMLElement | null>;
+  targetRef: React.RefObject<HTMLElement | null>;
+};
 
 const fetchLodgingList = async (pageParam: number) => {
   const response = await fetch(`/api/lodgings?offset=${pageParam}`);
@@ -6,7 +12,7 @@ const fetchLodgingList = async (pageParam: number) => {
   return result;
 };
 
-export function useInfiniteScroll() {
+export function useInfiniteScroll({ containerRef, targetRef }: UseInfiniteScrollProps) {
   const { data, error, fetchNextPage, hasNextPage, isFetchingNextPage } = useInfiniteQuery({
     queryKey: ['get-lodgings-infinite'],
     queryFn: ({ pageParam = 0 }) => fetchLodgingList(pageParam),
@@ -15,6 +21,24 @@ export function useInfiniteScroll() {
       return lastPage.hasNext ? lastPage.nextOffset : undefined;
     },
   });
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      const observer = new IntersectionObserver(
+        (entries) => {
+          if (entries[0].isIntersecting && hasNextPage && !isFetchingNextPage) {
+            fetchNextPage();
+          }
+        },
+        { root: containerRef.current, threshold: 0.8 }
+      );
+
+      if (targetRef.current) observer.observe(targetRef.current);
+      return () => observer.disconnect();
+    }, 500);
+
+    return () => clearTimeout(timer);
+  }, [targetRef.current, hasNextPage, fetchNextPage]);
 
   return { lodging: data, error, fetchNextPage, hasNextPage, isFetchingNextPage };
 }
